Initial version of autocc code
We can use a fairly trivial mechanism to detect from the process environment
which toolchain we need to call out to. If that fails we can inspect the filesystem
using the `PATH` variable.

We'll try to keep things as cheap as possible, and then we `execvpe` to replace the
running process with the real underlying compiler, while forcing the `argv[0]` name
to `/usr/bin/cc`.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>